% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{Illinois Counties.}
\format{A data frame with 7 variables
\describe{
\item{\code{name}}{County name}
\item{\code{id}}{ID number}
\item{\code{type}}{Type of the county: Cook, Collar, Rural, or Urban}
\item{\code{region}}{Region of the county: Central, North, or South}
\item{\code{circuit}}{Illinois juditical circuit number}
\item{\code{IDOC}}{Illinois Department of Correction code}
\item{\code{FIPS}}{Federal Information Processing Standards code}
}}
\usage{
counties
}
\description{
A \code{SpatialPolygonsDataFrame} object containing the Illinois counties polygons with
additional attributes. See "Format" below for more on the \code{data} table for
the attributes.
}
\keyword{datasets}
