% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fbicrime}
\alias{fbicrime}
\title{Crime in Illinois 2011-2015 from FBI.}
\format{A data frame with 13 variables:
\describe{
\item{\code{year}}{Calendar year}
\item{\code{county}}{County name}
\item{\code{violentCrime}}{Total sum of violent crime, including murder, rape, roberry and aggravated assault}
\item{\code{murder}}{Count of murder}
\item{\code{rape_old}}{Count of rape, based on an old definition, until 2013}
\item{\code{rape_new}}{Count of rape, based on a new definition, from 2014 onward}
\item{\code{robbery}}{Count of robbery}
\item{\code{aggAssault}}{Count of aggrevated assault}
\item{\code{propertyCrime}}{Total sum of property crime, including  burglary, larceny-theft, motor vehicle-theft and arson}
\item{\code{burglary}}{Count of burglary}
\item{\code{larcenyTft}}{Count of larceny-theft}
\item{\code{MVTft}}{Count of motor vehicle-theft}
\item{\code{arson}}{Count ofarson}
}}
\source{
\url{https://ucr.fbi.gov/ucr-publications}
}
\usage{
fbicrime
}
\description{
The county-level crime data in Illinois from 2011 to 2015 extracted from
the Federal Bureau of Investigation's Crime in the United States (CIUS)
Uniform Crime Reports (UCR) publications.
}
\keyword{datasets}
