% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ispcrime}
\alias{ispcrime}
\title{Crime in Illinois 2011-2015 from ISP.}
\format{A data frame with 12 variables:
\describe{
\item{\code{year}}{year}
\item{\code{county}}{county}
\item{\code{violentCrime}}{total sum of violent crime, including murder, rape, roberry and aggravated assault}
\item{\code{murder}}{count of murder}
\item{\code{rape}}{count of rape}
\item{\code{robbery}}{count of robbery}
\item{\code{aggAssault}}{count of aggrevated assault}
\item{\code{propertyCrime}}{total sum of property crime, including burglary, larceny-theft, motor vehicle-theft and arson}
\item{\code{burglary}}{count of burglary}
\item{\code{larcenyTft}}{count of larceny-theft}
\item{\code{MVTft}}{count of motor vehicle-theft}
\item{\code{arson}}{count ofarson}
}}
\source{
\url{https://www.isp.state.il.us/crime/ucrhome.cfm}
}
\usage{
ispcrime
}
\description{
The county-level crime data in Illinois from 2011 to 2015 extracted from
the Illinois State Police's Unifrom Crime Reports (UCR) annual reports.
}
\keyword{datasets}
