% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{Regional Categories for Illinois Counties.}
\format{A data frame with 2 variables: \code{region} and \code{county}}
\source{
\url{https://www.isp.state.il.us/crime/ucrhome.cfm}
}
\usage{
regions
}
\description{
Illinois counties are categorised into three larger regions: Northern,
Central, and Southern. Note that the Cook county is of its own category.
}
\keyword{datasets}
