% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ispcrime}
\alias{ispcrime}
\title{Crime in Illinois 2011-2015 from ISP.}
\format{A data frame with 12 variables:
\describe{
\item{\code{year}}{Calendar Year}
\item{\code{county}}{County name}
\item{\code{violentCrime}}{Total sum of violent crime, including murder, rape, roberry and aggravated assault}
\item{\code{murder}}{Count of murder}
\item{\code{rape}}{Count of rape}
\item{\code{robbery}}{Count of robbery}
\item{\code{aggAssault}}{Count of aggrevated assault}
\item{\code{propertyCrime}}{Total sum of property crime, including burglary, larceny-theft, motor vehicle-theft and arson}
\item{\code{burglary}}{Count of burglary}
\item{\code{larcenyTft}}{Count of larceny-theft}
\item{\code{MVTft}}{Count of motor vehicle-theft}
\item{\code{arson}}{Count ofarson}
}}
\source{
\url{https://www.isp.state.il.us/crime/ucrhome.cfm}
}
\usage{
ispcrime
}
\description{
The county-level crime data in Illinois from 2011 to 2015 extracted from
the Illinois State Police's Unifrom Crime Reports (UCR) annual reports.
}
\keyword{datasets}
